fix(package): max 20 octcts for serial
A certificate serial number is not decimal conforming. That is the bytes in a serial number do not necessarily map to a printable ASCII character.
eg: 0x00 is a valid serial number and can not be represented in a human readable format (atleast one that can be directly mapped to the ACSII table).

fix #84